alacritty: Improve update script

+++ update_alacritty.sh
@@ -1,9 +1,12 @@
 cd $1
 git fetch
 if [ -n "$(git diff)" ]; then
+    echo "Alacritty is already up-to-date."
     exit 0
 fi
 git merge FETCH_HEAD
 make app
-rm -r /Applications/Alacritty.app
-mv ./target/release/osx/Alacritty.app /Applications/Alacritty.app
+if [ $? -eq 0 ]; then
+    rm -r /Applications/Alacritty.app
+    mv ./target/release/osx/Alacritty.app /Applications/Alacritty.app
+fi
